refactor: mail로 이름변경 및 DB 데이터타입 수정
