+++ src/main/resources/db/migration/V1_8__Add_email_history.sql
@@ -1,10 +1,10 @@
-create table email_history
+create table mail_history
 (
     id         bigint not null auto_increment,
     subject    varchar(255) not null,
-    body       varchar(255) not null,
+    body       longtext not null,
     sender     varchar(255) not null,
-    recipients varchar(255) not null,
+    recipients longtext not null,
     sent_time  datetime(6) not null,
     primary key (id)
 ) engine = InnoDB
