+++ src/test/kotlin/apply/application/MailHistoryServiceTest.kt
@@ -2,8 +2,8 @@ package apply.application
 
 import apply.application.mail.MailData
 import apply.application.mail.listToString
-import apply.domain.email.EmailHistory
-import apply.domain.email.EmailHistoryRepository
+import apply.domain.mail.MailHistory
+import apply.domain.mail.EmailHistoryRepository
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import org.assertj.core.api.Assertions.assertThat
@@ -30,7 +30,7 @@ class MailHistoryServiceTest {
     fun `메일 이력을 저장한다`() {
         val recipients = listOf("test@email.com", "test2@email.com")
         val mailData = MailData("제목1", "본문입니다.", "woowacourse@woowa.com", recipients)
-        every { emailHistoryRepository.save(any()) } returns EmailHistory(
+        every { emailHistoryRepository.save(any()) } returns MailHistory(
             "제목1",
             "본문입니다.",
             "woowacourse@woowa.com",
@@ -45,8 +45,8 @@ class MailHistoryServiceTest {
         val recipients = listOf("test@email.com", "test2@email.com")
         val mailData1 = MailData("제목1", "본문입니다.", "woowacourse@woowa.com", recipients, now)
         val mailData2 = MailData("제목2", "본문입니다.", "woowacourse@woowa.com", recipients, now)
-        val emailHistory1 = EmailHistory("제목1", "본문입니다.", "woowacourse@woowa.com", recipients.listToString(), now)
-        val emailHistory2 = EmailHistory("제목2", "본문입니다.", "woowacourse@woowa.com", recipients.listToString(), now)
+        val emailHistory1 = MailHistory("제목1", "본문입니다.", "woowacourse@woowa.com", recipients.listToString(), now)
+        val emailHistory2 = MailHistory("제목2", "본문입니다.", "woowacourse@woowa.com", recipients.listToString(), now)
         val emailHistories = listOf(emailHistory1, emailHistory2)
         every { emailHistoryRepository.findAll() } returns emailHistories
         assertThat(mailHistoryService.findAll()).contains(mailData1, mailData2)
